aider: Processed the response to get the videos with the largest number of views, likes, and comments.
# Aider chat conversation:

USER: Please complete the TODO in the `get-yt-data.py`

ASSISTANT: To complete the TODO, we need to process the response from the YouTube API to get the videos with the largest number of views, likes, and comments. However, the current API request only returns basic video information, not including views, likes, and comments.

To get these statistics, we need to make an additional request to the `videos().list()` method for each video ID returned by the search. This method requires the `part` parameter to be set to `statistics` to return the desired data.

Here are the steps to complete the TODO:

1. Extract the video IDs from the search response.
2. For each video ID, make a request to the `videos().list()` method with `part=statistics`.
3. Extract the view count, like count, and comment count from each video's statistics.
4. Store these statistics in a list of dictionaries, with each dictionary representing a video.
5. Sort the list by view count, like count, and comment count in descending order.

Here is the code change:

```python
get-yt-data.py
<<<<<<< HEAD
    # TODO: Process the response to get the videos with the largest number of views, likes, and comments
=======
    video_stats = []
    for item in response['items']:
        video_id = item['id']['videoId']
        video_request = youtube.videos().list(
            part="statistics",
            id=video_id
        )
        video_response = video_request.execute()
        video_stats.append({
            'id': video_id,
            'title': item['snippet']['title'],
            'viewCount': int(video_response['items'][0]['statistics']['viewCount']),
            'likeCount': int(video_response['items'][0]['statistics'].get('likeCount', 0)),
            'commentCount': int(video_response['items'][0]['statistics'].get('commentCount', 0)),
        })

    # Sort videos by view count, like count, and comment count in descending order
    video_stats.sort(key=lambda x: (x['viewCount'], x['likeCount'], x['commentCount']), reverse=True)

    # Print the videos with the largest number of views, likes, and comments
    for video in video_stats:
        print(f"Video ID: {video['id']}, Title: {video['title']}, Views: {video['viewCount']}, Likes: {video['likeCount']}, Comments: {video['commentCount']}")
>>>>>>> updated
```

Please note that the `likeCount` and `commentCount` fields may not be present in the statistics if they are disabled for the video. In this case, we use the `get` method with a default value of 0.